Modify folder structure
