+++ src/Web/ClientApp/src/components/__tests__/Home.test.js
@@ -1,9 +1,8 @@
 import React from 'react';
-import ReactTestUtils from 'react-dom/test-utils';
-import { MemoryRouter } from 'react-router-dom';
 import { render, unmountComponentAtNode } from 'react-dom';
+import { MemoryRouter } from 'react-router-dom';
 import { act } from 'react-dom/test-utils';
-import { NavMenu } from '../components/NavMenu';
+import { Home } from '../Home';
 
 let container = null;
 beforeEach(() => {
@@ -21,20 +20,9 @@ it('renders without crashing', async () => {
   await act(async () => {
     render(
       <MemoryRouter>
-        <NavMenu />
+        <Home />
       </MemoryRouter>
       , container
     );
-  });
-
-  const button = document.getElementById('menu-icon-button');
-  await act(async () => {
-    button.click();
-  });
-
-  const menu = document.getElementById('menu-aria');
-  await act(async () => {
-    menu.click();
-    ReactTestUtils.Simulate.keyDown(menu, { key: "Enter", keyCode: 13, which: 13 })
   });
 });
