+++ src/Web/ClientApp/src/components/__tests__/NavMenu.test.js
@@ -1,8 +1,9 @@
 import React from 'react';
-import { render, unmountComponentAtNode } from 'react-dom';
+import ReactTestUtils from 'react-dom/test-utils';
 import { MemoryRouter } from 'react-router-dom';
+import { render, unmountComponentAtNode } from 'react-dom';
 import { act } from 'react-dom/test-utils';
-import { Details } from '../inspection/Details';
+import { NavMenu } from '../NavMenu';
 
 let container = null;
 beforeEach(() => {
@@ -20,11 +21,20 @@ it('renders without crashing', async () => {
   await act(async () => {
     render(
       <MemoryRouter>
-        <Details
-          match={{ params: { id: 'guid' } }}
-        />
+        <NavMenu />
       </MemoryRouter>
       , container
     );
+  });
+
+  const button = document.getElementById('menu-icon-button');
+  await act(async () => {
+    button.click();
+  });
+
+  const menu = document.getElementById('menu-aria');
+  await act(async () => {
+    menu.click();
+    ReactTestUtils.Simulate.keyDown(menu, { key: "Enter", keyCode: 13, which: 13 })
   });
 });
