+++ src/Web/ClientApp/src/inspection/__tests__/Create.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, unmountComponentAtNode } from 'react-dom';
 import { MemoryRouter } from 'react-router-dom';
 import { act } from 'react-dom/test-utils';
-import { ChoicesTemplate } from '../categories/ChoicesTemplate';
+import { Create } from '../Create';
 
 let container = null;
 beforeEach(() => {
@@ -17,25 +17,12 @@ afterEach(() => {
 });
 
 it('renders without crashing', async () => {
-  const templates = [
-    { choices: ['choice1-1', 'choice1-2'] },
-    { choices: ['choice2-1', 'choice2-2'] },
-    { choices: ['choice3-1', 'choice3-2'] },
-  ];
-  jest.spyOn(global, 'fetch').mockImplementation(() =>
-    Promise.resolve({
-      json: () => Promise.resolve(templates)
-    })
-  );
-
   await act(async () => {
     render(
       <MemoryRouter>
-        <ChoicesTemplate />
+        <Create />
       </MemoryRouter>
       , container
     );
   });
-
-  global.fetch.mockRestore();
 });
